+++ src/commands/encounters/tavern/restfulNight.ts
@@ -1,7 +1,7 @@
 import { CommandInteraction, MessageEmbed } from "discord.js";
+import { adjustHP } from "../../../character/adjustHP";
+import { awardXP } from "../../../character/awardXP";
 import { d6 } from "../../../gameState";
-import { adjustHP } from "../../../adjustHP";
-import { awardXP } from "../../../awardXP";
 
 export async function restfulNight(
   interaction: CommandInteraction
