+++ src/commands/equip.ts
@@ -5,10 +5,10 @@ import {
   MessageActionRow,
   MessageSelectMenu,
 } from "discord.js";
-import { equipItem } from "../equipItem";
 import { updateCharacter } from "../updateCharacter";
 import { getUserCharacter } from "../getUserCharacter";
-import { Item, itemEmbed, itemIsEquippable } from "../utils/equipment";
+import { Item, itemEmbed, itemIsEquippable } from "../equipment/equipment";
+import { equipItem } from "../character/equipItem";
 
 export const command = new SlashCommandBuilder()
   .setName("equip")
