+++ src/commands/renew.ts
@@ -1,10 +1,10 @@
 import { SlashCommandBuilder } from "@discordjs/builders";
 import { CommandInteraction, Message, MessageEmbed } from "discord.js";
 import { isCharacterOnCooldown } from "../character/isCharacterOnCooldown";
-import { adjustHP } from "../adjustHP";
 import { getUserCharacter } from "../getUserCharacter";
 import { cooldownRemainingText } from "../utils";
 import { hpBarField } from "../character/hpBar/hpBarField";
+import { adjustHP } from "../character/adjustHP";
 
 export const command = new SlashCommandBuilder()
   .setName("renew")
