+++ src/equipment/equipment.ts
@@ -1,10 +1,10 @@
 import { CommandInteraction, Message, MessageEmbed } from "discord.js";
 import inspect from "../commands/inspect";
-import { equipItem } from "../equipItem";
-import { equipItemRow } from "../equipment/equipItemRow";
+import { equipItemRow } from "./equipItemRow";
 import { updateCharacter } from "../updateCharacter";
 import { getUserCharacter } from "../getUserCharacter";
 import { StatModifier } from "../statusEffects/StatModifier";
+import { equipItem } from "../character/equipItem";
 
 export type Item = {
   type: "weapon" | "armor" | "shield";
