+++ src/equipment/grantCharacterItem.ts
@@ -1,5 +1,5 @@
 import { Character } from "../character/Character";
-import { Item } from "../utils/equipment";
+import { Item } from "./equipment";
 
 export const grantCharacterItem = (
   character: Character,
